requirements.txt
gpiozero
pytelegraf
pyyaml
RPi.GPIO
psutil